requirements.txt
adal==1.2.5
applicationinsights==0.11.9
azure-cognitiveservices-vision-computervision==0.7.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.26
azure-core==1.10.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==12.0.0
azure-mgmt-storage==11.2.0
azureml-automl-core==1.20.0
azureml-core==1.20.0
azureml-dataprep==2.7.3
azureml-dataprep-native==27.0.0
azureml-dataprep-rslex==1.5.0
azureml-dataset-runtime==1.20.0
azureml-pipeline==1.20.0
azureml-pipeline-core==1.20.0
azureml-pipeline-steps==1.20.0
azureml-sdk==1.20.0
azureml-telemetry==1.20.0
azureml-train==1.20.0
azureml-train-automl-client==1.20.0
azureml-train-core==1.20.0
azureml-train-restclients-hyperdrive==1.20.0
backports.tempfile==1.0
backports.weakref==1.0.post1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
contextlib2==0.6.0.post1
cryptography==3.3.1
cycler==0.10.0
distro==1.5.0
docker==4.4.1
dotnetcore2==2.1.20
Flask==1.0.3
fusepy==3.0.1
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
jeepney==0.6.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==1.5.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
msal==1.8.0
msal-extensions==0.2.2
msrest==0.6.19
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.16.2
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.1.0
portalocker==1.7.1
pyarrow==1.0.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.4.1
SecretStorage==3.3.0
six==1.15.0
urllib3==1.26.2
websocket-client==0.57.0
Werkzeug==1.0.1
